refactorized footer component
